feat(money): integer minor-units migration + planning-feature review fixes (#73)
* perf(parser): switch untrusted regex matching to RE2J to eliminate ReDoS

Bank-statement parsing matches regex patterns from ParserConfig (synced
Firestore→Room) that are effectively untrusted — AI-generated and user
candidate patterns. java.util.regex / kotlin.text.Regex are NOT
interruptible, so withTimeout+runInterruptible in ParseStatementUseCase
cancels the coroutine while the worker thread keeps burning CPU on a
catastrophic-backtracking pattern. The string-truncation + ReDoS-heuristic
palliatives only narrowed the window.

Move the hot matching path in RegexStatementParser onto RE2 (com.google.re2j,
1.8), which matches in guaranteed linear time with no backtracking, removing
ReDoS as a class for this path:
- transaction_pattern, skip_patterns → RE2; line_fixups → RE2 where possible.
- RE2 uses Python (?P<name>) named-group syntax; configs store Java (?<name>).
  Convert at compile time inside the parser (normalizeNamedGroups), without
  mutating what's stored in the config/Firestore. Lookbehind (?<= / (?<! is
  left untouched. Documented as the single source of truth.
- Rewrite matchToTransaction for the RE2J Matcher API (group(int)/group(String),
  null-coalesced), preserving safeNamedGroup and positional/named behavior.

Incompatibility policy (RE2 has no lookaround/backreferences):
- incompatible transaction_pattern → empty result → triggers AI fallback;
- uncompilable skip_pattern → literal Pattern.quote fallback (preserves the
  long-standing "invalid skip pattern still filters its literal text");
- incompatible line_fixup → ReDoS-guarded kotlin.Regex fallback. Fixups are
  pure text-repair replace() ops and several real configs legitimately need
  lookaround (Freedom "Сумма в …обработке" reorder; Eurasian misplaced-time
  tempered-greedy token). Each fixup still passes RegexValidator and input is
  capped at MAX_LINE_LENGTH, so residual backtracking is bounded.

RegexValidator and the length cap stay as defense-in-depth. Adds tests: a
catastrophic pattern on long input completes under a hard timeout (linearity
proof), lookahead transaction/skip/fixup handling, named-group normalization,
and RE2-compilation of all five reference bank patterns. Only the 19
pre-existing *BankIntegrationTest (missing Firestore configs) still fail.

* test(parser): seed bank parser configs in a test-only fixture for integration tests

The 5 *BankIntegrationTest suites (kaspi/freedom/forte/bereke/eurasian) assert
real transaction counts and first-row values against the checked-in sample PDFs,
but RegexParserProfileTestFactory loaded default_parser_config.json — which ships
empty ({"banks": []}) — so BankDetector.detect() returned null and every test
died on `!!`. They had never been runnable offline (failed identically on main);
this is unrelated to the RE2J migration.

Fix WITHOUT changing what the app ships: the app deliberately bundles an empty
default_parser_config.json and resolves real configs from Firestore at runtime
(never hard-wired to a bundled bank set). So the reference configs live in a new
TEST-ONLY resource, test_bank_configs.json, loaded only by
RegexParserProfileTestFactory and never packaged into the APK. The bundled
production (and mirrored test) default_parser_config.json stay empty, so
RegexParserProfileSyncTest still passes.

The 5 reference profiles were reconstructed for testing from:
- freedom/eurasian patterns informed by the live Firestore parser_candidates;
- kaspi/forte/bereke from the calibrated inline configs in the sibling
  *BankParserTest classes;
- validated against the real PDFs until each assertion matched (kaspi=34,
  forte=80, bereke=163, etc.).

Two calibration details surfaced against the real Bereke statement:
- dropped the over-generic markers "Freedom Bank"/"ForteBank", which matched
  those banks' names appearing as payee text inside the Bereke PDF and caused
  misdetection; the BIC/domain markers uniquely identify each bank;
- added "Cash withdrawal from ATM" and "Fee" to the Bereke operation alternation
  (the 7 rows missing to reach 163).

core:parser: 262 tests, 0 failures (was 19), 7 pre-existing PdfTableExtractor
JVM skips. Production bundled config unchanged (empty).

* feat(money): integer minor-units migration + planning-feature review fixes

Large-scale change across the whole app.

Money model (spec 019):
- Store/compute all monetary amounts as Long minor units (fixed scale 2),
  never Double; conversion helpers in core:model (money/). Room schema v9
  + Migration_8_9; Firestore dual-write (legacy major + encrypted *Minor,
  read *Minor-first); uniqueHash reconstructs legacy amount string for
  cross-version dedup stability.

Stage-8 planning-feature review fixes (recurring / budgets / debts):
- Debts: debt-payment transaction now updates the account balance and is
  reversed (balance + soft-delete) when the payment/debt is deleted; edit
  no longer duplicates the linked transaction; block total < paid with a
  clear error; surface save/delete failures via Snackbar; payment date
  picker; DebtDao.observeById.
- Recurring: re-anchor DatePicker UTC->local (off-by-one fix); deterministic
  uniqueHash + insertOrIgnore so generation is idempotent / dedup-safe across
  devices; YEARLY preserves Feb 29; first occurrence honours dayOfMonth/
  dayOfWeek; catch-up cap; sensible default day pickers.
- Budgets: exclude already-budgeted categories and surface save errors;
  revive soft-deleted budget instead of orphaning its Firestore doc; pure
  mapper; rounded percent.
- Across: replace raw hex with theme tokens; new localized strings.

Tests: GeneratePendingTransactionsUseCase date math, DebtPaymentRepositoryImpl
balance apply/reverse, BudgetRepositoryImpl revive.

* test(import): disambiguate insertOrIgnore overload in ImportTransactionsUseCaseTest

Adding the single-entity insertOrIgnore overload to TransactionDao made the
test's untyped any() ambiguous; specify any<List<TransactionEntity>>().

* fix(ui): hide swipe-to-delete icon at rest so it no longer overlaps row content

The swipe-to-delete backgroundContent drew the trash icon unconditionally;
only the background color animated with the swipe. At rest (transparent bg)
the opaque icon stayed painted at CenterEnd, and in dark theme its onError
tint is dark red — so it overlapped the amount/time on the right of every
row. Gate the icon behind isSwiping, matching the debts screens.

+++ core/database/src/main/java/com/atelbay/money_manager/core/database/entity/BudgetEntity.kt
@@ -23,7 +23,7 @@ import androidx.room.PrimaryKey
 data class BudgetEntity(
     @PrimaryKey(autoGenerate = true) val id: Long = 0,
     val categoryId: Long,
-    val monthlyLimit: Double,
+    val monthlyLimit: Long,
     val createdAt: Long,
     val remoteId: String? = null,
     val updatedAt: Long = 0,
